refactor: generalize ProcessTile
