+++ tasklist/client/src/common/processes/index.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+import type {ProcessDefinition} from '@vzeta/camunda-api-zod-schemas/operate';
+import type {Process} from 'v1/api/types';
+
+type MultiModeProcess = Process | ProcessDefinition;
+
+export type {MultiModeProcess};
